Removing polling for resetting video playback
Previously a Handler would check once a second if the video has
exceed the range selector's duration. This is now changed to
use a timer. When the video is paused the timer is deleted
until the video playback is resumed.

As an additional simplification, the playback location is no longer
saved when the video is paused; instead playback always restarts
from the beginning.