Make progress indeterminate until FFmpeg status updates start
If the video is being cut and the start of the output is
not the start of the file, FFmpeg needs to work its way through
the input file until it starts encoding. During this time the
progress bar will now be indeterminate, signaling that the
encoding has not yet started.