Process FFmpeg calls in a JobService
Managing the FFmpeg calls in the Activity leaves open the
possibility that the Activity is killed due to memory
constraints, leaving the FFmpeg call orphaned or killed.
To manage this and other situations better, the managing
is now moved to a JobService.

The JobService is responsible for receiving encoding requests,
reporting status, and reporting the final result. Presently
only one file can be processed at a time.

If the Activity goes away while a file is being processed, when
the Activity is created again it will determine if a file
is being processed in the background. If it is, the progress
bar will be shown with the cancel button.