Initialize FFmpeg in onCreate()
The FFmpegProcessService will attempt to initialize FFmpeg,
but only after it is asked to process a file. This means
that when the first file is selected its media info will
fail to be queried. Instead of moving everything to the
service, simply initialize in onCreate(), and post the
unsupported message of initialization fails.