Show video select button after video cancel
If the video were canceled when the app was brought back
and encoding was not processing in the backgroud but there
was a job to process, then the cancel callback will never
occur. To let the buttons for video selection to be shown
in this case, unconditionally show them after a cancel.