Use embedded file picker to select video file
Determining the selected file's path using an arbitrary activity
was not reliable. On some devices and with some activities the
path was not being returned, or worse was resulting in a crash.

To simplify the process of selecting a media file, the
Storage Chooser is used. This handles displaying a file selection
activity and returns the result as an absolute file path.