Use no nonsense file picker
The codekidX file picker had some issues which occasionally
resulted in crashes. Some of these bugs have been reported
to the upstream project, but as the bug reports are from
Google Play not that much information was available to give.

The no nonsense file picker is used by the NewPipe app, and
it appears to be successful. Switching to it to hopefully
simplify the file selection process and make it more robust.