Use ACRA to capture crashes, easily send via email
A number of crashes have been recorded in Google Play, but
it is tricky to determine what is going on. The expectation is
that if crash logs can be captured by users directly, they
may contain more data. In addition, users may enter additional
details which may help track down issues. There will also be
the ability to contact the users for additional info.

Note that ProGuard was necessary, as inclduing ACRA resulted in
a apk which had more methods than the limit of 64k.