Skip palette generation step when converting to gif
GIF files only support 256 colors. Previously a pass over the input
video was taken to develop a good color palette before converting
to gif. However, it was found that on a recent FFmpeg version that
the palette generation step was taking a significant amount of
memory and time to complete. For non-trivial videos it was not
possible to create a gif.

This change removes the palette generation step. This will likely
cause the color quality of the output gif to decrease, however
it should be much quicker to generate the gif.