Handle SEND intents for encoding files
This enables the app to handle SEND intents, which send
data to the app for processing.

ffmpeg, which encodes the files, must process a file path
instead of a stream. Because of this, the data in the intent
must be saved locally first then staged for processing.
As soon as the SEND intent is received the data is copied over,
then a preview posted to the user.

There are a few aspects which are not yet resolved, such as
posting a permissions dialog if the app does not have the permissions
yet, and the sent file's name is not preserved so the output
name makes sense.