Restructure permission checks to callback on success
Previously there was only one reason for a permission check,
namely to select a video. However, there are now ENCODE and
SEND intent which need to be handled. To reduce the complexity
of handling what is requesting the permission, a callback
is now being handed to getPermission() which will invoke
it if the read/write permissions are granted.