Override output file name on SEND intent
Previously the file name of the input file was used
as the basis of the output file name. However, because
the file from a SEND intent is saved to a tmp file, its
name is meaningless. This change will preserve the name
so it can be used to label the output file.