Add setting for file search location
When open the file browser to find a file, the starting location
used to be external storage. It was mentioned that the last used
path would be easier, so this was instead made the default.

It turns out that navigating to the internal storage device
may prevent one from navigating back to external storage. The
way to recover was to wipe the app's data, resetting its last
used path. To avoid this, a new setting is added which sets what
the start path should be. The default is to use the last used
path, however it can be fixed to the external storage location.