Set the 'preset' setting for h264 to 'faster'
An article examined the time/quality trade-off for h264 and
concluded that the best default option was 'faster'. Compared
to 'medium', which is the default for FFmpeg, 'faster' reduced
encoding times ~73% and reduced the quality in a likely
imperceptible manner.

This sounds like a good trade-off generally, especially for a
mobile device where power usage is a concern. This commit
changes the 'preset' setting to 'faster'.

The article was found here:
https://streaminglearningcenter.com/blogs/saving-encoding-adjust-encoding-configuration-increase-capacity.html