Lower the min sdk to 19
The min sdk was increased to 21 because JobService was being used.
Since, that has been removed, so the app should work on sdk 19.
The sdk could be lower yet, as the app did compile on sdk 16. However,
unless someone is interested in trying that out on a device, sdk 19
is fine for now.