Do not stub out info for invalid files
Originally it was not known if parsing a media file
would generally succeed, and as a hedge some sane defaults
were given to the user if something went wrong.

It seems that parsing media files works well. If an
invalid file is provided, the video playback will fail
(which is expected), but encoding options are still provided.
In this case, the sane defaults are actually more confusing.
Removing them, as they are not really needed.