Fix detection of encode complete in test
The -d argument was thought to delete the existing Android
logs. However, this was not the case. As a result, the
python tests were concluding that the file being encoded
finished near immediately, and concluded that the incomplete
file existed so the test passed.

This change uses the -t argument of logcat to gather all logs
since a given time, and will pass in the time the test started.
This way, the python test will properly wait until the encoding
has completed before pulling the file. Further, it will not
conclude that encoding was successful if it never even started.