Check for alt. name when converting to Audio/Video Codec
ffprobe sometimes returns a different name for a codec than
ffmpeg accepts as arguments. An example is 'libvpx-vp9' vs 'VP9'.
Because of this, sometimes when a media file is selected its
codec(s) may not be recognized.

This commit adds a pretty name to AudioCodec, and uses the
pretty name when looking up a video and audio codec.