Prevent re-attempting video preview playback on error
Previously if a media file was selected but the VideoView
could not play it, the VideoView would popup an error dialog
and at the end of a timeout the video preview was started again.
The result was an endless stream of error dialogs.

This error case is now detected, and if a file fails to
play a Toast is shown and the video playback is stopped.